reorganises file structure
